requirements.txt
aiocryptopay==0.4.8
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.10.0
async-timeout==5.0.1
attrs==25.3.0
bcrypt==5.0.0
certifi==2023.11.17
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
cryptography==46.0.2
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
exceptiongroup==1.3.0
fastapi==0.117.1
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
httplib2==0.20.2
idna==3.10
importlib-metadata==4.6.4
invoke==2.2.0
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
magic-filter==1.0.12
more-itertools==8.10.0
multidict==6.6.4
ngrok==1.4.0
oauthlib==3.2.0
paramiko==4.0.0
propcache==0.3.2
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
PyGObject==3.42.1
PyJWT==2.3.0
PyNaCl==1.6.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu4
python-dotenv==1.1.1
Pillow==11.0.0
PyYAML==5.4.1
qrcode==8.2
requests==2.32.5
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.43
ssh-import-id==5.11
starlette==0.48.0
StrEnum==0.4.15
systemd-python==234
typing-inspection==0.4.1
typing_extensions==4.15.0
ubuntu-drivers-common==0.0.0
unattended-upgrades==0.1
urllib3==2.5.0
uvicorn==0.36.0
wadllib==1.3.6
xkit==0.0.0
yarl==1.20.1
zipp==1.0.0
